--- src/main/java/com/barbel/memberserver/global/utill/MemberUpdateDTOMapper.java
@@ -1,17 +1,0 @@
-package com.barbel.memberserver.global.utill;
-
-import com.barbel.memberserver.domain.member.dto.MemberUpdateRequest;
-import com.barbel.memberserver.domain.member.document.Member;
-
-public class MemberUpdateDTOMapper {
-  public static Member memberUpdateRequestToMember(MemberUpdateRequest memberUpdateRequest, Member member) {
-    member.setNickname(memberUpdateRequest.getNickname());
-    member.setAddress(memberUpdateRequest.getAddress());
-    member.setAddressDetail(memberUpdateRequest.getAddressDetail());
-    member.setIntroduce(memberUpdateRequest.getIntroduce());
-    member.setMajors(memberUpdateRequest.getMajors());
-    member.setInterests(memberUpdateRequest.getInterests());
-    member.setRole(memberUpdateRequest.getRole());
-    return member;
-  }
-}
